Moved plotting routines to plotting directory
