+++ blimpy/filterbank.py
@@ -26,14 +26,13 @@ import sys
 import six
 
 from astropy.time import Time
-import scipy.stats
-from matplotlib.ticker import NullFormatter
 
 import logging as logger
 
 try:
     from .utils import db, lin, rebin, closest, unpack_2to8
     from .sigproc import *
+    from .plotting import *
 except:
     from utils import db, lin, rebin, closest, unpack_2to8
     from sigproc import *
@@ -52,26 +51,13 @@ except ImportError:
 
 #import pdb #pdb.set_trace()
 
-# Check if $DISPLAY is set (for handling plotting on remote machines with no X-forwarding)
-if 'DISPLAY' in os.environ.keys():
-    import pylab as plt
-else:
-    import matplotlib
-    matplotlib.use('Agg')
-    import pylab as plt
-
-plt.rcParams['axes.formatter.useoffset'] = False
-
 
 ###
 # Config values
 ###
 
-MAX_PLT_POINTS      = 65536                  # Max number of points in matplotlib plot
-MAX_IMSHOW_POINTS   = (8192, 4096)           # Max number of points in imshow plot
 MAX_DATA_ARRAY_SIZE = 1024 * 1024 * 1024 * 2 # Max size of data array to load into memory
 MAX_HEADER_BLOCKS   = 100                    # Max size of header (in 512-byte blocks)
-
 
 # Telescope coordinates (needed for LSR calc)
 parkes_coords = (-32.998370, 148.263659,  324.00)
@@ -506,409 +492,13 @@ class Filterbank(object):
         plot_t_end  = self.timestamps[-1] + (self.timestamps[1] - self.timestamps[0])
 
         if MJD_time:
-            extent=(plot_f_begin, plot_f_begin_end, plot_t_begin, plot_t_end)
+            extent=(plot_f_begin, plot_f_end, plot_t_begin, plot_t_end)
         else:
             extent=(plot_f_begin, plot_f_end, 0.0,(plot_t_end-plot_t_begin)*24.*60.*60)
 
         return extent
 
-    def plot_spectrum(self, t=0, f_start=None, f_stop=None, logged=False, if_id=0, c=None, **kwargs):
-        """ Plot frequency spectrum of a given file
 
-        Args:
-            t (int): integration number to plot (0 -> len(data))
-            logged (bool): Plot in linear (False) or dB units (True)
-            if_id (int): IF identification (if multiple IF signals in file)
-            c: color for line
-            kwargs: keyword args to be passed to matplotlib plot()
-        """
-        if self.header[b'nbits'] <=2:
-            logged = False
-            t='all'
-        ax = plt.gca()
-
-        plot_f, plot_data = self.grab_data(f_start, f_stop, if_id)
-
-        #Using accending frequency for all plots.
-        if self.header[b'foff'] < 0:
-            plot_data = plot_data[..., ::-1] # Reverse data
-            plot_f = plot_f[::-1]
-
-        if isinstance(t, int):
-            print("extracting integration %i..." % t)
-            plot_data = plot_data[t]
-        elif t == b'all':
-            print("averaging along time axis...")
-            #Since the data has been squeezed, the axis for time goes away if only one bin, causing a bug with axis=1
-            if len(plot_data.shape) > 1:
-                plot_data = plot_data.mean(axis=0)
-            else:
-                plot_data = plot_data.mean()
-        else:
-            raise RuntimeError("Unknown integration %s" % t)
-
-        # Rebin to max number of points
-        dec_fac_x = 1
-        if plot_data.shape[0] > MAX_PLT_POINTS:
-            dec_fac_x = int(plot_data.shape[0] / MAX_PLT_POINTS)
-
-        plot_data = rebin(plot_data, dec_fac_x, 1)
-        plot_f    = rebin(plot_f, dec_fac_x, 1)
-
-        if not c:
-            kwargs['c'] = '#333333'
-
-        if logged:
-            plt.plot(plot_f, db(plot_data),label='Stokes I', **kwargs)
-            plt.ylabel("Power [dB]")
-        else:
-
-            plt.plot(plot_f, plot_data,label='Stokes I', **kwargs)
-            plt.ylabel("Power [counts]")
-        plt.xlabel("Frequency [MHz]")
-        plt.legend()
-
-        try:
-            plt.title(self.header[b'source_name'])
-        except KeyError:
-            plt.title(self.filename)
-
-        plt.xlim(plot_f[0], plot_f[-1])
-
-    def plot_spectrum_min_max(self, t=0, f_start=None, f_stop=None, logged=False, if_id=0, c=None, **kwargs):
-        """ Plot frequency spectrum of a given file
-
-        Args:
-            logged (bool): Plot in linear (False) or dB units (True)
-            if_id (int): IF identification (if multiple IF signals in file)
-            c: color for line
-            kwargs: keyword args to be passed to matplotlib plot()
-        """
-        ax = plt.gca()
-
-        plot_f, plot_data = self.grab_data(f_start, f_stop, if_id)
-
-        #Using accending frequency for all plots.
-        if self.header[b'foff'] < 0:
-            plot_data = plot_data[..., ::-1] # Reverse data
-            plot_f = plot_f[::-1]
-
-        fig_max = plot_data[0].max()
-        fig_min = plot_data[0].min()
-
-        print("averaging along time axis...")
-
-        #Since the data has been squeezed, the axis for time goes away if only one bin, causing a bug with axis=1
-        if len(plot_data.shape) > 1:
-            plot_max = plot_data.max(axis=0)
-            plot_min = plot_data.min(axis=0)
-            plot_data = plot_data.mean(axis=0)
-        else:
-            plot_max = plot_data.max()
-            plot_min = plot_data.min()
-            plot_data = plot_data.mean()
-
-        # Rebin to max number of points
-        dec_fac_x = 1
-        MAX_PLT_POINTS = 8*64  # Low resoluition to see the difference.
-        if plot_data.shape[0] > MAX_PLT_POINTS:
-            dec_fac_x = int(plot_data.shape[0] / MAX_PLT_POINTS)
-
-        plot_data = rebin(plot_data, dec_fac_x, 1)
-        plot_min = rebin(plot_min, dec_fac_x, 1)
-        plot_max = rebin(plot_max, dec_fac_x, 1)
-        plot_f    = rebin(plot_f, dec_fac_x, 1)
-
-        if logged:
-            plt.plot(plot_f, db(plot_data), "#333333", label='mean', **kwargs)
-            plt.plot(plot_f, db(plot_max),  "#e74c3c", label='max', **kwargs)
-            plt.plot(plot_f, db(plot_min),  '#3b5b92', label='min', **kwargs)
-            plt.ylabel("Power [dB]")
-        else:
-            plt.plot(plot_f, plot_data,  "#333333", label='mean', **kwargs)
-            plt.plot(plot_f, plot_max,   "#e74c3c", label='max', **kwargs)
-            plt.plot(plot_f, plot_min,   '#3b5b92', label='min', **kwargs)
-            plt.ylabel("Power [counts]")
-        plt.xlabel("Frequency [MHz]")
-        plt.legend()
-
-        try:
-            plt.title(self.header[b'source_name'])
-        except KeyError:
-            plt.title(self.filename)
-
-        plt.xlim(plot_f[0], plot_f[-1])
-        if logged:
-            plt.ylim(db(fig_min),db(fig_max))
-
-    def plot_waterfall(self, f_start=None, f_stop=None, if_id=0, logged=True, cb=True, MJD_time=False, **kwargs):
-        """ Plot waterfall of data
-
-        Args:
-            f_start (float): start frequency, in MHz
-            f_stop (float): stop frequency, in MHz
-            logged (bool): Plot in linear (False) or dB units (True),
-            cb (bool): for plotting the colorbar
-            kwargs: keyword args to be passed to matplotlib imshow()
-        """
-
-        plot_f, plot_data = self.grab_data(f_start, f_stop, if_id)
-
-        #Using accending frequency for all plots.
-        if self.header[b'foff'] < 0:
-            plot_data = plot_data[..., ::-1] # Reverse data
-            plot_f = plot_f[::-1]
-
-        if logged:
-            plot_data = db(plot_data)
-
-        # Make sure waterfall plot is under 4k*4k
-        dec_fac_x, dec_fac_y = 1, 1
-        if plot_data.shape[0] > MAX_IMSHOW_POINTS[0]:
-            dec_fac_x = int(plot_data.shape[0] / MAX_IMSHOW_POINTS[0])
-
-        if plot_data.shape[1] > MAX_IMSHOW_POINTS[1]:
-            dec_fac_y =  int(plot_data.shape[1] /  MAX_IMSHOW_POINTS[1])
-
-        plot_data = rebin(plot_data, dec_fac_x, dec_fac_y)
-
-        try:
-            plt.title(self.header[b'source_name'])
-        except KeyError:
-            plt.title(self.filename)
-
-        extent = self._calc_extent(plot_f=plot_f,plot_t=self.timestamps,MJD_time=MJD_time)
-
-        plt.imshow(plot_data,
-            aspect='auto',
-            origin='lower',
-            rasterized=True,
-            interpolation='nearest',
-            extent=extent,
-            cmap='viridis',
-            **kwargs
-        )
-        if cb:
-            plt.colorbar()
-        plt.xlabel("Frequency [MHz]")
-        if MJD_time:
-            plt.ylabel("Time [MJD]")
-        else:
-            plt.ylabel("Time [s]")
-
-    def plot_time_series(self, f_start=None, f_stop=None, if_id=0, logged=True, orientation='h', MJD_time=False, **kwargs):
-        """ Plot the time series.
-
-         Args:
-            f_start (float): start frequency, in MHz
-            f_stop (float): stop frequency, in MHz
-            logged (bool): Plot in linear (False) or dB units (True),
-            kwargs: keyword args to be passed to matplotlib imshow()
-        """
-
-        ax = plt.gca()
-        plot_f, plot_data = self.grab_data(f_start, f_stop, if_id)
-
-        if logged and self.header[b'nbits'] >= 8:
-            plot_data = db(plot_data)
-
-        #Since the data has been squeezed, the axis for time goes away if only one bin, causing a bug with axis=1
-        if len(plot_data.shape) > 1:
-            plot_data = plot_data.mean(axis=1)
-        else:
-            plot_data = plot_data.mean()
-
-        #Make proper time axis for plotting (but only for plotting!). Note that this makes the values inclusive.
-        extent = self._calc_extent(plot_f=plot_f,plot_t=self.timestamps,MJD_time=MJD_time)
-        plot_t = np.linspace(extent[2],extent[3],len(self.timestamps))
-
-        if MJD_time:
-            tlabel = "Time [MJD]"
-        else:
-            tlabel = "Time [s]"
-
-        if logged:
-            plabel = "Power [dB]"
-        else:
-            plabel = "Power [counts]"
-
-        # Reverse oder if vertical orientation.
-        if 'v' in orientation:
-            plt.plot(plot_data, plot_t, **kwargs)
-            plt.xlabel(plabel)
-
-        else:
-            plt.plot(plot_t, plot_data, **kwargs)
-            plt.xlabel(tlabel)
-            plt.ylabel(plabel)
-
-        ax.autoscale(axis='both',tight=True)
-
-    def plot_kurtosis(self, f_start=None, f_stop=None, if_id=0, **kwargs):
-        """ Plot kurtosis
-
-         Args:
-            f_start (float): start frequency, in MHz
-            f_stop (float): stop frequency, in MHz
-            kwargs: keyword args to be passed to matplotlib imshow()
-        """
-        ax = plt.gca()
-
-        plot_f, plot_data = self.grab_data(f_start, f_stop, if_id)
-
-        #Using accending frequency for all plots.
-        if self.header[b'foff'] < 0:
-            plot_data = plot_data[..., ::-1] # Reverse data
-            plot_f = plot_f[::-1]
-
-        try:
-            plot_kurtosis = scipy.stats.kurtosis(plot_data, axis=0, nan_policy='omit')
-        except:
-            plot_kurtosis = plot_data*0.0
-
-        plt.plot(plot_f, plot_kurtosis, **kwargs)
-        plt.ylabel("Kurtosis")
-        plt.xlabel("Frequency [MHz]")
-
-        plt.xlim(plot_f[0], plot_f[-1])
-
-    def plot_all(self, t=0, f_start=None, f_stop=None, logged=False, if_id=0, kurtosis=True, **kwargs):
-        """ Plot waterfall of data as well as spectrum; also, placeholder to make even more complicated plots in the future.
-
-        Args:
-            f_start (float): start frequency, in MHz
-            f_stop (float): stop frequency, in MHz
-            logged (bool): Plot in linear (False) or dB units (True),
-            t (int): integration number to plot (0 -> len(data))
-            logged (bool): Plot in linear (False) or dB units (True)
-            if_id (int): IF identification (if multiple IF signals in file)
-            kwargs: keyword args to be passed to matplotlib plot() and imshow()
-        """
-        if self.header[b'nbits'] <=2:
-            logged = False
-
-        nullfmt = NullFormatter()  # no labels
-
-        # definitions for the axes
-        left, width = 0.35, 0.5
-        bottom, height = 0.45, 0.5
-        width2, height2 = 0.1125, 0.15
-        bottom2, left2 = bottom - height2 - .025, left - width2 - .02
-        bottom3, left3 = bottom2 - height2 - .025, 0.075
-
-        rect_waterfall = [left, bottom, width, height]
-        rect_colorbar = [left + width, bottom, .025, height]
-        rect_spectrum = [left, bottom2, width, height2]
-        rect_min_max = [left, bottom3, width, height2]
-        rect_timeseries = [left + width, bottom, width2, height]
-        rect_kurtosis = [left3, bottom3, 0.25, height2]
-        rect_header = [left3 - .05, bottom, 0.2, height]
-
-        # --------
-        #         axColorbar = plt.axes(rect_colorbar)
-        #         print 'Ploting Colorbar'
-        #         print plot_data.max()
-        #         print plot_data.min()
-        #
-        #         plot_colorbar = range(plot_data.min(),plot_data.max(),int((plot_data.max()-plot_data.min())/plot_data.shape[0]))
-        #         plot_colorbar = np.array([[plot_colorbar],[plot_colorbar]])
-        #
-        #         plt.imshow(plot_colorbar,aspect='auto', rasterized=True, interpolation='nearest',)
-
-        #         axColorbar.xaxis.set_major_formatter(nullfmt)
-        #         axColorbar.yaxis.set_major_formatter(nullfmt)
-
-        #         heatmap = axColorbar.pcolor(plot_data, edgecolors = 'none', picker=True)
-        #         plt.colorbar(heatmap, cax = axColorbar)
-
-
-        # --------
-        axMinMax = plt.axes(rect_min_max)
-        print('Plotting Min Max')
-        self.plot_spectrum_min_max(logged=logged, f_start=f_start, f_stop=f_stop, t=t)
-        plt.title('')
-        axMinMax.yaxis.tick_right()
-        axMinMax.yaxis.set_label_position("right")
-
-        # --------
-        axSpectrum = plt.axes(rect_spectrum,sharex=axMinMax)
-        print('Plotting Spectrum')
-        self.plot_spectrum(logged=logged, f_start=f_start, f_stop=f_stop, t=t)
-        plt.title('')
-        axSpectrum.yaxis.tick_right()
-        axSpectrum.yaxis.set_label_position("right")
-        plt.xlabel('')
-#        axSpectrum.xaxis.set_major_formatter(nullfmt)
-        plt.setp(axSpectrum.get_xticklabels(), visible=False)
-
-        # --------
-        axWaterfall = plt.axes(rect_waterfall,sharex=axMinMax)
-        print('Plotting Waterfall')
-        self.plot_waterfall(f_start=f_start, f_stop=f_stop, logged=logged, cb=False)
-        plt.xlabel('')
-
-        # no labels
-#        axWaterfall.xaxis.set_major_formatter(nullfmt)
-        plt.setp(axWaterfall.get_xticklabels(), visible=False)
-
-        # --------
-        axTimeseries = plt.axes(rect_timeseries)
-        print('Plotting Timeseries')
-        self.plot_time_series(f_start=f_start, f_stop=f_stop, orientation='v')
-        axTimeseries.yaxis.set_major_formatter(nullfmt)
-#        axTimeseries.xaxis.set_major_formatter(nullfmt)
-
-        # --------
-        # Could exclude since it takes much longer to run than the other plots.
-        if kurtosis:
-            axKurtosis = plt.axes(rect_kurtosis)
-            print('Plotting Kurtosis')
-            self.plot_kurtosis(f_start=f_start, f_stop=f_stop)
-
-
-        # --------
-        axHeader = plt.axes(rect_header)
-        print('Plotting Header')
-        # Generate nicer header
-        telescopes = {0: 'Fake data',
-                      1: 'Arecibo',
-                      2: 'Ooty',
-                      3: 'Nancay',
-                      4: 'Parkes',
-                      5: 'Jodrell',
-                      6: 'GBT',
-                      8: 'Effelsberg',
-                      10: 'SRT',
-                      64: 'MeerKAT',
-                      65: 'KAT7'
-                      }
-
-        telescope = telescopes.get(self.header[b"telescope_id"], self.header[b"telescope_id"])
-
-        plot_header = "%14s: %s\n" % ("TELESCOPE_ID", telescope)
-        for key in (b'SRC_RAJ', b'SRC_DEJ', b'TSTART', b'NCHANS', b'NBEAMS', b'NIFS', b'NBITS'):
-            try:
-                plot_header += "%14s: %s\n" % (key, self.header[key.lower()])
-            except KeyError:
-                pass
-        fch1 = "%6.6f MHz" % self.header[b'fch1']
-
-        foff = (self.header[b'foff'] * 1e6 * u.Hz)
-        if np.abs(foff) > 1e6 * u.Hz:
-            foff = str(foff.to('MHz'))
-        elif np.abs(foff) > 1e3 * u.Hz:
-            foff = str(foff.to('kHz'))
-        else:
-            foff = str(foff.to('Hz'))
-
-        plot_header += "%14s: %s\n" % ("FCH1", fch1)
-        plot_header += "%14s: %s\n" % ("FOFF", foff)
-
-        plt.text(0.05, .95, plot_header, ha='left', va='top', wrap=True)
-
-        axHeader.set_facecolor('white')
-        axHeader.xaxis.set_major_formatter(nullfmt)
-        axHeader.yaxis.set_major_formatter(nullfmt)
 
     def write_to_filterbank(self, filename_out):
         """ Write data to blimpy file.
